Move QueryParams to apicast dir
So it can be used from other policies, and not just from the URL
rewriting one.
